scripts/skills/handwriting_recognition/requirements.txt
paddleocr>=2.7.0
pytesseract>=0.3.10
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0
torch>=2.0.0
tensorflow>=2.13.0
